requirements.txt
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.5
colorama==0.4.5
comtypes==1.1.11
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
Cython==0.29.30
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dlib @ file:///D:/bld/dlib_1617441818780/work
easyocr==1.5.0
editdistance==0.6.0
efficientnet==1.0.0
entrypoints==0.4
essential-generators==1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==1.12
fonttools==4.27.1
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
graphviz==0.8.4
grpcio==1.34.1
gTTS==2.2.4
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.6
imageio==2.15.0
imgaug==0.4.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
playsound==1.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyclipper==1.3.0.post3
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.1.1
pywin32==304
pywinpty==1.1.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.18.4
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.8.0
Shapely==1.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.2.0
SpeechRecognition==3.8.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.22
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
